refactor(Atsumaru): rename provider dirs and update imports
