+++ src/Atsumaru/implementations/manga-details-providing/main.ts
