+++ src/Atsumaru/implementations/manga-details-providing/parsers.ts
