+++ src/Atsumaru/implementations/search-results-providing/parsers.ts
